+++ api/batch_dubbing_music.sh
@@ -60,6 +60,7 @@ else
 	if [[ $COUNT -gt 0 ]] ; then
 		VIDFILES=`find input/vr/dubbing/music -maxdepth 1 -type f -name '*.mp4' -o -name '*.webm'`
 		for nextinputfile in $VIDFILES ; do
+			[ -e "$nextinputfile" ] || continue
 			[ -e user/default/comfyui_stereoscopic/.pipelinepause ] && exit 0
 			INDEX+=1
 			echo "$INDEX/$COUNT" >input/vr/dubbing/music/BATCHPROGRESS.TXT
